+++ R/plotPresGrid.R
@@ -14,6 +14,9 @@
 #' @param type one of either "presence" or "density". If "density", then boxes will be colored according
 #'   to the number of detections in each \code{timeBin} will be plotted. If "presence", then each box
 #'   will be colored by \code{fill}
+#' @param by (optional) if not \code{NULL}, specifies the name of a column in \code{x} to split and
+#'   color the rectangles by. Only valid for presence plots.
+#' @param alpha opacity of rectangles, only used if \code{by} is not \code{NULL}
 #' @param gps (optional) if not \code{NULL}, a data.frame of GPS coordinates covering the date range
 #'   of \code{x}. These are used to calculate sunrise and sunset information which is shown as a
 #'   shaded dark region in the background of the plot. The data.frame must have columns "UTC",
@@ -33,7 +36,8 @@
 #' @examples
 #'
 #' df <- data.frame(UTC = as.POSIXct(runif(1e2, min=0, max=7*24*3600),
-#'                                   origin='1970-01-01 00:00:00', tz='UTC'))
+#'                                   origin='1970-01-01 00:00:00', tz='UTC'),
+#'                  label = sample(letters[1:3], 1e2, replace=TRUE))
 #' plotPresGrid(df, type='presence', bin='hour')
 #' plotPresGrid(df, type='density', bin='hour')
 #' plotPresGrid(df, type='density', bin='30min')
@@ -41,23 +45,34 @@
 #'                  Latitude=32.4,
 #'                  Longitude = -118)
 #' plotPresGrid(df, gps=gps, bin='hour')
+#' # coloring presence grid by label column
+#' plotPresGrid(df, gps=gps, by='label')
+#' # can be confusing if there is a lot of overlap, ggplot output can be split
+#' library(ggplot2)
+#' plotPresGrid(df, gps=gps, by='label') + facet_wrap(vars(label), ncol=2)
+#' # using "by" with type="density" defaults to this facet_wrap behavior
+#' # since color is already being used to represent density
+#' plotPresGrid(df, gps=gps, by='label', type='density')
+#' # can adjust facet_wrap parameters by adding it again
+#' plotPresGrid(df, gps=gps, by='label', type='density') + facet_wrap(vars(label), ncol=2)
 #'
 #' @export
 #'
 #' @importFrom lubridate floor_date hour period minute
 #' @importFrom suncalc getSunlightTimes
-#' @importFrom data.table setDT setDF setkeyv data.table
-#' @importFrom viridisLite viridis
+#' @importFrom scales viridis_pal
 #' @import ggplot2
 #' @import dplyr
 #'
-plotPresGrid <- function(x, start=NULL, end=NULL, 
+plotPresGrid <- function(x, start=NULL, end=NULL,
                          bin=c('hour', 'minute', '30min', '15min'),
-                         type=c('presence', 'density'), gps=NULL,
+                         type=c('presence', 'density'),
+                         by=NULL, alpha=.5,
+                         gps=NULL,
                          format = c('%m/%d/%Y %H:%M:%S', '%m-%d-%Y %H:%M:%S',
                                     '%Y/%m/%d %H:%M:%S', '%Y-%m-%d %H:%M:%S'),
                          fill='blue', color=NA,
-                         cmap=viridis(25), title=TRUE) {
+                         cmap=viridis_pal()(25), title=TRUE) {
     if(!'UTC' %in% colnames(x)) {
         stop('"x" must have column UTC')
     }
@@ -65,7 +80,7 @@ plotPresGrid <- function(x, start=NULL, end=NULL,
        is.factor(x$UTC)) {
         x$UTC <- parseToUTC(as.character(x$UTC), format=format, tz='UTC')
     }
-    
+
     if(is.null(start)) {
         start <- floor_date(min(x$UTC), unit='day')
     }
@@ -109,20 +124,56 @@ plotPresGrid <- function(x, start=NULL, end=NULL,
            }
     )
     x$day <- floor_date(x$UTC, 'day')
+
     type <- match.arg(type)
     if(type == 'presence') {
-        x <- select(x, all_of(c('plot_min', 'plot_max', 'day'))) %>%
-            distinct()
-        g <- g +
-            geom_rect(data=x, aes(xmin=.data$plot_min, xmax=.data$plot_max, ymin=.data$day, ymax=.data$day + period(1, 'day')), fill=fill, col=color)
+        if(is.null(by)) {
+            x <- select(x, all_of(c('plot_min', 'plot_max', 'day'))) %>%
+                distinct()
+            g <- g +
+                geom_rect(data=x, aes(xmin=.data$plot_min,
+                                      xmax=.data$plot_max,
+                                      ymin=.data$day,
+                                      ymax=.data$day + period(1, 'day')),
+                          fill=fill, col=color)
+        } else {
+            x <- select(x, all_of(c('plot_min', 'plot_max', 'day', by))) %>%
+                distinct()
+
+            g <- g +
+                geom_rect(data=x, aes(xmin=.data$plot_min,
+                                      xmax=.data$plot_max,
+                                      ymin=.data$day,
+                                      ymax=.data$day + period(1, 'day'),
+                                      fill=.data[[by]]),
+                          alpha=alpha, col=color)
+        }
     }
     if(type == 'density') {
-        x <- group_by(x, .data$day, .data$plot_min, .data$plot_max) %>%
-            summarise(count=n(), .groups='keep')
-        g <- g +
-            geom_rect(data=x, aes(xmin=.data$plot_min, xmax=.data$plot_max, ymin=.data$day, ymax=.data$day + period(1, 'day'), fill=count)) +
-            scale_fill_gradientn(colors=cmap)
-        tlab <- paste0('Calls/', tlab)
+        if(is.null(by)) {
+            x <- group_by(x, .data$day, .data$plot_min, .data$plot_max) %>%
+                summarise(count=n(), .groups='keep')
+            g <- g +
+                geom_rect(data=x, aes(xmin=.data$plot_min,
+                                      xmax=.data$plot_max,
+                                      ymin=.data$day,
+                                      ymax=.data$day + period(1, 'day'),
+                                      fill=count)) +
+                scale_fill_gradientn(colors=cmap)
+            tlab <- paste0('Calls/', tlab)
+        } else {
+            x <- group_by(x, .data$day, .data$plot_min, .data$plot_max, .data[[by]]) %>%
+                summarise(count=n(), .groups='keep')
+            g <- g +
+                geom_rect(data=x, aes(xmin=.data$plot_min,
+                                      xmax=.data$plot_max,
+                                      ymin=.data$day,
+                                      ymax=.data$day + period(1, 'day'),
+                                      fill=count)) +
+                scale_fill_gradientn(colors=cmap) +
+                facet_wrap(by)
+            tlab <- paste0('Calls/', tlab)
+        }
     }
     if(isTRUE(title)) {
         title <- paste0('Call ', oneUp(type), ' (', tlab, ')')
@@ -130,18 +181,38 @@ plotPresGrid <- function(x, start=NULL, end=NULL,
     g + ggtitle(title)
 }
 
+globalVariables('closest')
+
 makeSunriseDf <- function(start, end, gps) {
     gpsCols <- c('UTC', 'Latitude', 'Longitude')
     if(!all(gpsCols %in% colnames(gps))) {
         stop('gps must have columns UTC, Latitude, and Longitude')
     }
-    gps <- gps[gpsCols]
-    sun_df <- data.table(UTC = seq(from=start, to=end, 'day'))
-    setkeyv(sun_df, 'UTC')
-    setDT(gps)
-    setkeyv(gps, 'UTC')
-    sun_df <- gps[sun_df, roll='nearest']
-    setDF(sun_df)
+    gps <- distinct(gps[gpsCols])
+    ###
+    gps$sUTC <- gps$UTC
+    x <- data.frame(UTC = seq(from=start, to=end, 'day'))
+    x$isGreater <- x$UTC >= min(gps$UTC)
+
+    x <- bind_rows(lapply(split(x, x$isGreater), function(d) {
+        if(d$isGreater[1]) {
+            d <- left_join(d, gps[c('sUTC', 'Latitude', 'Longitude')], join_by(closest('UTC' >= 'sUTC')))
+        } else {
+            d <- left_join(d, gps[c('sUTC', 'Latitude', 'Longitude')], join_by(closest('UTC' <= 'sUTC')))
+        }
+        d
+    }))
+    x$sUTC <- NULL
+    x$isGreater <- NULL
+    gps$sUTC <- NULL
+    sun_df <- x
+    ###
+    # sun_df <- data.table(UTC = seq(from=start, to=end, 'day'))
+    # setkeyv(sun_df, 'UTC')
+    # setDT(gps)
+    # setkeyv(gps, 'UTC')
+    # sun_df <- gps[sun_df, roll='nearest']
+    # setDF(sun_df)
     sun_df$Date <- as.Date(sun_df$UTC)
     sun_df$sunrise <- NA
     sun_df$sunset <- NA